Refactoring in preparation for core merge (#131)
* Split up the Shiny_Updates class and move some functions around
* Get rid of `su_update_plugin()`
* Improve JavaScript
