+++ src/ajax-actions.php
@@ -2,6 +2,8 @@
 /**
  * Ajax callbacks for Shiny Updates.
  *
+ * @todo Merge: Add to wp-admin/includes/ajax-actions.php
+ *
  * @package Shiny_Updates
  */
 
@@ -23,7 +25,7 @@ function wp_ajax_install_theme() {
 
 	$status = array(
 		'install' => 'theme',
-		'slug'    => sanitize_key( $_POST['slug'] ),
+		'slug'    => sanitize_key( wp_unslash( $_POST['slug'] ) ),
 	);
 
 	if ( ! current_user_can( 'install_themes' ) ) {
@@ -90,7 +92,7 @@ function wp_ajax_update_theme() {
 		) );
 	}
 
-	$stylesheet = sanitize_key( $_POST['slug'] );
+	$stylesheet = sanitize_key( wp_unslash( $_POST['slug'] ) );
 	$status     = array(
 		'update'     => 'theme',
 		'slug'       => $stylesheet,
@@ -172,7 +174,7 @@ function wp_ajax_delete_theme() {
 		) );
 	}
 
-	$stylesheet = sanitize_key( $_POST['slug'] );
+	$stylesheet = sanitize_key( wp_unslash( $_POST['slug'] ) );
 	$status     = array(
 		'delete' => 'theme',
 		'slug'   => $stylesheet,
@@ -237,7 +239,7 @@ function wp_ajax_install_plugin() {
 
 	$status = array(
 		'install' => 'plugin',
-		'slug'    => sanitize_key( $_POST['slug'] ),
+		'slug'    => sanitize_key( wp_unslash( $_POST['slug'] ) ),
 	);
 
 	if ( ! current_user_can( 'install_plugins' ) ) {
@@ -249,7 +251,7 @@ function wp_ajax_install_plugin() {
 	include_once( ABSPATH . 'wp-admin/includes/plugin-install.php' );
 
 	$api = plugins_api( 'plugin_information', array(
-		'slug'   => sanitize_key( $_POST['slug'] ),
+		'slug'   => sanitize_key( wp_unslash( $_POST['slug'] ) ),
 		'fields' => array(
 			'sections' => false,
 		),
@@ -292,6 +294,8 @@ function wp_ajax_install_plugin() {
 
 /**
  * AJAX handler for updating a plugin.
+ *
+ * @todo Merge: Replace wp_ajax_update_plugin()
  *
  * @since 4.2.0
  *
